Fix namespaces not being properly added to arrays.
When the SOAP request body is passed as a hash, one may specify the
value for a key to be an array of hashes. In this case, each hash
in the array is treated as being contained within a element specified by
the key. The namespace applied to the elements specified within the
array was previously the default namespace. This fix recursively
determines the proper namespaces for these elements.

- Modifiy lib/savon/soap/xml.rb to recursively determine proper
  namespaces for elements contained within arrays
- Modifiy spec/fixtures/wsdl/multiple_namespaces.xml to add a new
  operation that properly exhibits the problem
- Modifiy spec/savon/client_spec.rb to add a new test that exhibits the
  problem
- Modifiy spec/savon/client_spec.rb to change an old test that failed
  after the change. The failure was not a regression. In this case, the
  same namespace was applied but with a different namespace name.