Raise an error if Savon::Multipart is not available
Don't try to load it, assume that the end user will do this - this way
we don't need to potentially call require multiple times during runtime.

Also fixed the multipart specs so they'd be less likely to interfere with
other ones and make those fail. Instead of mucking with the spec-level
globals object, create one for multipart when we need it.