introduced the OperationBuilder
OperationBuilder replaces the old operation handling which was not
built with multiple documents in mind. it uses the new section methods
introduced for each Document and aggregated by the DocumentCollection.

bye bye, LegacyOperationParser!

with this commit we're also finally dropping the snakecase Symbol
conversion for operation names. from here on out we just use the
original names all throughout Wasabi.