Make tests pass under rubinius
A little surprised by the extent of gems that are not included in
the standard lib, but that's the state of the rubinius world. It
seems that rubinius is still difficult to maintain compatibility
with, but I've noticed (and been told before) that rubinius does help
expose bugs that other rubies will cover up, so I'm willing to personally
invest this time.