Added WSDL fixture for no_message_tag test.
Added test case for :no_message_tag global option.

This option appears to be a workaround for a bug
in either the wsdl parser or the building of the XML
response message. I am not entirely sure what the issue
is but the :no_message_tag option solves the problem
I ran into when using savon with the 3PLCentral XML/Soap
client APIs.