Fix an unintended crash in Builder 2.1.2
It's an older version of Builder, but if you're using Rails 3.0.20, several
gems (including ActionPack and ActiveModel) depend on this. This doesn't
seem too scary a change to make, but I'm not entirely sure of the repercussions
of this. However, this is the default value for the attribute, so I think this
is probably safe.

Fixes #582