Loosen versions of dependencies
It has been rather annoying and cumbersome to update these as
new patch versions of gyoku, wasabi, etc are released. My feeling
is that limiting them to the minor version number is fine (because
we might want to verify that a feature exists, for example). This means
that we only need to change this when the minor version changes