Allow :any IN expected hash as value
This to accomodate for example random (or hard to mock) values in the message.

In our case this was needed when the request/message needed a random RequestID.

In theory we would be able to mock this random value, on the other hand we needed to introduce a mocking library in our test sutie just for this.