Add support for rpc encoded wsdl
After Wasabi add support for multi part on messages, now it's returning a hash
with parts, and Savon needs to handle correctly.

So on builder, instead of just add a single message to body, now handle if it's
a hash and build the body containing the respective parts, with its type and
arguments passed via `locals` variable.

This way we add support for builder create a message that contains multi parts.