Only treat top level fault nodes as a SOAPFault.
When a response contains a nested fault node (anywhere other than a direct child of `Body`), the `SOAPFault.present?` returns true, but Nori is unable to properly find the fault node (`nori.find(to_hash, 'Fault')`). In that scenario, you are left w/ a `NoMethodError: undefined method 'each' for nil:NilClass` (spec added to show this scenario).

Rather than treat this scenario as a SOAPFault, it would be preferable to parse the response normally. This will also cover the scenarios of potentially having other node names that end in `Fault>`, `Code>`, `Reason>`, etc.

This should address the following open issues:
https://github.com/savonrb/savon/issues/841
https://github.com/savonrb/savon/issues/769