Only treat top level fault nodes as a SOAPFault
When a response contains a nested fault node (anywhere other than a direct child of `Body`), the `SOAPFault.present?` class method returns true, but Nori is unable to properly find the fault node (`nori.find(to_hash, 'Fault')`). In that scenario, you are left w/ a `NoMethodError: undefined method 'each' for nil:NilClass` (spec added to show this scenario).

Rather than treat this scenario as a `SOAPFault`, it would be preferable to parse the response normally. This will also cover the scenarios of potentially having other node names that end in `Fault>`, `Code>`, `Reason>`, etc.

This should address the following open issues:
https://github.com/savonrb/savon/issues/841
https://github.com/savonrb/savon/issues/769

Note: Using a regex matcher here caused benchmark tests of the `SOAPFault.present?` method to run ~50% slower; running the `.present?` spec 5 million times for another_soap_fault.xml took 20 seconds when using `.include?("Fault>")`, and 30 seconds when using the regex matcher, so we're talking about a difference of a couple microsecond for parsing a relatively small file.