Allow the write timeout option to be set
Write timeout is supported by HTTPI but not by Savon.

While not having the setting is limiting, it can also cause issues in
adapters. For example, when setting timeouts per operation in HTTP, any
timeouts not supplied will be set to 0.25 seconds which can lead to
unexpected surprises for users