Pass the provided Savon/custom logger to Faraday
If the Savon logger isn't passed to Faraday, it builds a new logger to $stdout and logs to it.
This results in stdout logging whenever `log: true` is set as a savon option, instead of logging to a file or so.

Passing the Savon/custom logger to Faraday avoids the creation of a second logger object by Faraday and allows to capture all logs to a user provided Logger object.

This patch also removes the `level` option, which is no recognized Faraday option.
There is the option `log_level` only, which determines which log level which the messages are flagged by Faraday.
In contrast the `log_level` option of Savon determines the minimum level of logs, which are captured by the logger.